refactor : move GameQuizDetail QUZ-75
